backend: move sources under src/, add events, entities, users and tests
Restructures the C++ backend into src/game, src/transport and src/users.
Adds the user manager: libsodium Argon2 password hashing, with unknown
accounts verified against a dummy hash so login timing does not reveal
which emails exist. The user repository is still a stub.

+++ src/backend-cpp/src/transport/codec.cpp
@@ -34,7 +34,7 @@ std::string Codec::Encode(const GameManager &game_manager) const {
     for (auto &&[id, pose]: game_manager.GetSlimesWithPoses()) {
         const SlimeRepublics::Vec2 vec(pose.position.x, pose.position.y);
         const SlimeRepublics::Direction heading = ToWire(pose.heading);
-        v.push_back(SlimeRepublics::CreateSlime(builder, id, &vec, heading));
+        v.push_back(SlimeRepublics::CreateSlime(builder, (uint32_t)id, &vec, heading));
     }
 
     const auto svo = builder.CreateVector(v);
